pyproject.toml
[tool.poetry]
name = "circuitgraph"
version = "0.1.0"
description = ""
authors = ["Vojtěch Sýkora [vojtasykor@gmail.com]"]
readme = "README.md"
packages = [
    { include = "converter" },
    { include = "extraction" }
]

[tool.poetry.dependencies]
python = "^3.11,<3.13"
decorator = "4.4.2"
lxml = "^5.3.0"
networkx = "^3.4.2"
numpy = "<2.0"
pillow = "^11.1.0"
pyparsing = "3.1.0"
matplotlib = "^3.10.0"
flask = "^3.1.0"
requests = "^2.32.3"
rdflib = "6.2.0"
levenshtein = "^0.26.1"
torchmetrics = {extras = ["detection"], version = "^1.6.1"}
torchinfo = "^1.8.0"
black = "^25.1.0"
isort = "^6.0.1"
tqdm = "^4.67.1"
bs4 = "^0.0.2"
selenium = "^4.29.0"
scikit-image = "^0.25.2"
pyqt5 = "^5.15.11"
docling = "^2.25.2"
opencv-python = "^4.11.0.86"
ipykernel = "^6.29.5"
tueplots = "^0.2.0"
psutil = "^7.0.0"
torchvision = "^0.22.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"